remove soda_plugin_mars:is_providing()
is_providing() is redundant since there is a specific publishing
request, so SODA will know in beforehand all valid requests (which is
actually those that have been published). Any requests for content for
any such file is expected to return OK, and the fetch should raise a
SodaException otherwise.

And if we have a need for allowing several SODA plugins to coexist,
it's probably better to configure SODA to use certain plugins for
certain files/queries.