Remove Celery worker registrar
Simplify code reasoning by removing the registrar worker and
moving the registration logic into the scheduler worker
instead. This avoids any issues that can occur mostly when a
request is being registered concurrently while the schedule_tasks
is being run.

Generate a request uuid to the user in the Flask app and defer db
registration to the scheduler worker using the requeust uuid as
the celery task id. This allows the request to be traced even
before it has been registered in the db.