Disable setting task status to REGISTERED
The scheduler join_sizing_tasks could sometimes wait indefinitely for
the result from a task although the result is already ready. Why this
happens is unclear, but it seems that it does not occur anymore if we
disable the setting to REGISTERED after the task is created.

Note that this means that we cannot distinguish between known/unknown
tasks until they have a corresponding DownloadRequest in the db.